requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
pillow==10.4.0
requests==2.32.3
Werkzeug==3.0.3
xmltodict==0.13.0
